+++ src/main/java/thedayoftoday/domain/auth/controller/AuthController.java
@@ -4,39 +4,35 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import thedayoftoday.domain.auth.dto.EmailCondeValidationDto;
-import thedayoftoday.domain.auth.dto.SendCodeRequestDto;
+import thedayoftoday.domain.auth.dto.LoginRequestDto;
 import thedayoftoday.domain.auth.dto.SignupRequestDto;
 import thedayoftoday.domain.auth.service.AuthService;
+import thedayoftoday.domain.auth.service.LoginService;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @RestController
-@RequestMapping("/user") //프론트 고치면 "/auth"로 고칠 예정
+//@RequestMapping("/auth") //프론트 고치면 "/auth"로 고칠 예정
 @RequiredArgsConstructor
 public class AuthController {
 
     private final AuthService authService;
+    private final LoginService loginService;
 
-    @PostMapping("/signup")
+    @PostMapping("/user/signup")
     public ResponseEntity<String> signup(@Valid @RequestBody SignupRequestDto signupRequestDto) {
         authService.join(signupRequestDto);
         return ResponseEntity.ok("회원가입이 완료되었습니다.");
     }
 
-    @GetMapping("/find-email")
-    public ResponseEntity<String> checkEmail(String email) {
-        authService.checkEmailExists(email);
-        return ResponseEntity.ok("사용 가능한 이메일입니다.");
-    }
+    @PostMapping("/swagger-auth/login")
+    public ResponseEntity<Map<String, String>> login(@RequestBody LoginRequestDto loginDto) {
+        String accessToken = loginService.login(loginDto);
 
-    @PostMapping("/send-code")
-    public ResponseEntity<String> sendCode(@RequestBody SendCodeRequestDto sendCodeRequestDto) {
-        authService.sendVerificationCode(sendCodeRequestDto);
-        return ResponseEntity.ok("인증 코드가 발송되었습니다.");
-    }
+        Map<String, String> response = new HashMap<>();
+        response.put("accessToken", accessToken);
 
-    @PostMapping("/check-code")
-    public ResponseEntity<String> checkCode(@RequestBody EmailCondeValidationDto emailCondeValidationDto) {
-        authService.verifyCode(emailCondeValidationDto);
-        return ResponseEntity.ok("정상적으로 인증되었습니다.");
+        return ResponseEntity.ok(response);
     }
 }
