--- src/main/java/thedayoftoday/domain/auth/dto/EmailCondeValidationDto.java
@@ -1,4 +1,0 @@
-package thedayoftoday.domain.auth.dto;
-
-public record EmailCondeValidationDto(String email, String code) {
-}
