--- src/main/java/thedayoftoday/domain/auth/dto/SendCodeRequestDto.java
@@ -1,4 +1,0 @@
-package thedayoftoday.domain.auth.dto;
-
-public record SendCodeRequestDto(String email) {
-}
