--- src/main/java/thedayoftoday/domain/auth/exception/EmailCodeExpireException.java
@@ -1,7 +1,0 @@
-package thedayoftoday.domain.auth.exception;
-
-public class EmailCodeExpireException extends RuntimeException {
-    public EmailCodeExpireException(String message) {
-        super(message);
-    }
-}
