--- src/main/java/thedayoftoday/domain/auth/exception/EmailCodeNotMatchException.java
@@ -1,7 +1,0 @@
-package thedayoftoday.domain.auth.exception;
-
-public class EmailCodeNotMatchException extends RuntimeException {
-  public EmailCodeNotMatchException(String message) {
-    super(message);
-  }
-}
