--- src/main/java/thedayoftoday/domain/auth/exception/EmailDuplicationException.java
@@ -1,7 +1,0 @@
-package thedayoftoday.domain.auth.exception;
-
-public class EmailDuplicationException extends RuntimeException {
-    public EmailDuplicationException(String message) {
-        super(message);
-    }
-}
