--- src/main/java/thedayoftoday/domain/auth/exception/ErrorCodeMessage.java
@@ -1,4 +1,0 @@
-package thedayoftoday.domain.auth.exception;
-
-public record ErrorCodeMessage(int code, String message) {
-}
