--- src/main/java/thedayoftoday/domain/auth/exception/MailSendException.java
@@ -1,7 +1,0 @@
-package thedayoftoday.domain.auth.exception;
-
-public class MailSendException extends RuntimeException {
-  public MailSendException(String message) {
-    super(message);
-  }
-}
