+++ src/main/java/thedayoftoday/domain/auth/mail/EmailCondeValidationDto.java
@@ -1,0 +1,4 @@
+package thedayoftoday.domain.auth.mail;
+
+public record EmailCondeValidationDto(String email, String code) {
+}
