+++ src/main/java/thedayoftoday/domain/auth/mail/SendCodeRequestDto.java
@@ -1,0 +1,4 @@
+package thedayoftoday.domain.auth.mail;
+
+public record SendCodeRequestDto(String email) {
+}
