+++ src/main/java/thedayoftoday/domain/auth/mail/exception/EmailCodeExpireException.java
@@ -1,0 +1,7 @@
+package thedayoftoday.domain.auth.mail.exception;
+
+public class EmailCodeExpireException extends RuntimeException {
+    public EmailCodeExpireException(String message) {
+        super(message);
+    }
+}
