+++ src/main/java/thedayoftoday/domain/auth/mail/exception/EmailCodeNotMatchException.java
@@ -1,0 +1,7 @@
+package thedayoftoday.domain.auth.mail.exception;
+
+public class EmailCodeNotMatchException extends RuntimeException {
+  public EmailCodeNotMatchException(String message) {
+    super(message);
+  }
+}
