+++ src/main/java/thedayoftoday/domain/auth/mail/exception/EmailDuplicationException.java
@@ -1,0 +1,7 @@
+package thedayoftoday.domain.auth.mail.exception;
+
+public class EmailDuplicationException extends RuntimeException {
+    public EmailDuplicationException(String message) {
+        super(message);
+    }
+}
