+++ src/main/java/thedayoftoday/domain/auth/mail/exception/MailSendException.java
@@ -1,0 +1,7 @@
+package thedayoftoday.domain.auth.mail.exception;
+
+public class MailSendException extends RuntimeException {
+  public MailSendException(String message) {
+    super(message);
+  }
+}
