+++ src/main/java/thedayoftoday/domain/auth/security/CorsProperties.java
@@ -1,0 +1,21 @@
+package thedayoftoday.domain.auth.security;
+
+import java.util.List;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "cors")
+public class CorsProperties {
+    private List<String> allowedOrigins;
+    private List<String> allowedMethods;
+    private List<String> allowedHeaders;
+    private Boolean allowCredentials;
+    private Long maxAge;
+    private List<String> exposedHeaders;
+}
