+++ src/main/java/thedayoftoday/domain/auth/security/entity/RefreshEntity.java
@@ -1,0 +1,21 @@
+package thedayoftoday.domain.auth.security.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class RefreshEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String email;
+    private String refresh;
+    private String expiration;
+}
