+++ src/main/java/thedayoftoday/domain/auth/security/repository/RefreshRepository.java
@@ -1,0 +1,13 @@
+package thedayoftoday.domain.auth.security.repository;
+
+import thedayoftoday.domain.auth.security.entity.RefreshEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface RefreshRepository extends JpaRepository<RefreshEntity, Long> {
+
+    Boolean existsByRefresh(String refresh);
+
+    @Transactional
+    void deleteByRefresh(String refresh);
+}
