+++ src/main/java/thedayoftoday/domain/user/controller/UserController.java
@@ -5,7 +5,7 @@ import thedayoftoday.domain.user.service.UserService;
 import thedayoftoday.domain.user.dto.PasswordUpdateRequest;
 import thedayoftoday.domain.user.dto.ResetPasswordRequestDto;
 import thedayoftoday.domain.user.dto.UserInfoDto;
-import thedayoftoday.security.CustomUserDetails;
+import thedayoftoday.domain.auth.security.CustomUserDetails;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
