+++ src/main/java/thedayoftoday/domain/user/service/UserService.java
@@ -7,9 +7,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import thedayoftoday.domain.auth.dto.SignupRequestDto;
-import thedayoftoday.domain.auth.exception.EmailDuplicationException;
+import thedayoftoday.domain.auth.mail.exception.EmailDuplicationException;
 import thedayoftoday.domain.auth.exception.PhoneNumberDuplicationExceptiono;
-import thedayoftoday.domain.user.entity.RoleType;
 import thedayoftoday.domain.user.repository.UserRepository;
 import thedayoftoday.domain.user.dto.PasswordUpdateRequest;
 import thedayoftoday.domain.user.dto.ResetPasswordRequestDto;
