+++ src/main/java/thedayoftoday/global/ErrorCodeMessage.java
@@ -1,0 +1,4 @@
+package thedayoftoday.global;
+
+public record ErrorCodeMessage(int code, String message) {
+}
