--- src/main/java/thedayoftoday/security/CustomUserDetails.java
@@ -1,68 +1,0 @@
-package thedayoftoday.security;
-
-import thedayoftoday.domain.user.entity.User;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-@Getter
-public class CustomUserDetails implements UserDetails {
-
-    private final Long userId;
-    private final String email;
-    private final String password;
-    private final String role;
-
-    public CustomUserDetails(User user) {
-        this.userId = user.getUserId();
-        this.email = user.getEmail();
-        this.password = user.getPassword();
-        this.role = user.getRole().toString();
-    }
-
-    public CustomUserDetails(String email, String role, Long userId) {
-        this.userId = userId;
-        this.email = email;
-        this.password = null;
-        this.role = role;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(role));
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
