--- src/main/java/thedayoftoday/security/entity/RefreshEntity.java
@@ -1,21 +1,0 @@
-package thedayoftoday.security.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class RefreshEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String email;
-    private String refresh;
-    private String expiration;
-}
