--- src/main/java/thedayoftoday/security/filter/JWTFilter.java
@@ -1,72 +1,0 @@
-package thedayoftoday.security.filter;
-
-import thedayoftoday.security.CustomUserDetails;
-import thedayoftoday.security.util.JWTUtil;
-import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-public class JWTFilter extends OncePerRequestFilter {
-
-    private final JWTUtil jwtUtil;
-
-    public JWTFilter(JWTUtil jwtUtil) {
-        this.jwtUtil = jwtUtil;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
-        String authorizationHeader = request.getHeader("Authorization");
-
-        if (authorizationHeader == null || !authorizationHeader.startsWith("Bearer ")) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        String accessToken = authorizationHeader.substring(7);
-
-        if (accessToken.trim().isEmpty()) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.getWriter().print("access token expired");
-            return;
-        }
-
-        String category = jwtUtil.getCategory(accessToken);
-        if (!"access".equals(category)) {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.getWriter().print("invalid access token");
-            return;
-        }
-
-        String email = jwtUtil.getUsername(accessToken);
-        String role = jwtUtil.getRole(accessToken);
-        Long userId = jwtUtil.getUserId(accessToken);
-
-        CustomUserDetails customUserDetails = new CustomUserDetails(email, role, userId);
-
-        Authentication authToken = new UsernamePasswordAuthenticationToken(
-                customUserDetails, null, customUserDetails.getAuthorities()
-        );
-        SecurityContextHolder.getContext().setAuthentication(authToken);
-        System.out.println("[JWTFilter] Set authentication with userId = " + userId);
-
-        filterChain.doFilter(request, response);
-    }
-}
