--- src/main/java/thedayoftoday/security/repository/RefreshRepository.java
@@ -1,13 +1,0 @@
-package thedayoftoday.security.repository;
-
-import thedayoftoday.security.entity.RefreshEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.transaction.annotation.Transactional;
-
-public interface RefreshRepository extends JpaRepository<RefreshEntity, Long> {
-
-    Boolean existsByRefresh(String refresh);
-
-    @Transactional
-    void deleteByRefresh(String refresh);
-}
